Move AFO builder traits to all_for_one module
