+++ crates/relayer-cosmos/src/base/all_for_one/builder.rs
@@ -1,5 +1,5 @@
 use async_trait::async_trait;
-use ibc_relayer_framework::base::builder::traits::birelay::CanBuildAfoBaseBiRelay;
+use ibc_relayer_framework::base::all_for_one::builder::CanBuildAfoBaseBiRelay;
 use ibc_relayer_framework::base::core::traits::error::HasErrorType;
 use ibc_relayer_framework::base::one_for_all::traits::birelay::OfaBiRelayPreset;
 use ibc_relayer_framework::base::one_for_all::types::birelay::OfaBiRelayWrapper;
