+++ crates/relayer-framework/src/base/all_for_one/mod.rs
@@ -3,6 +3,7 @@
 */
 
 pub mod birelay;
+pub mod builder;
 pub mod chain;
 pub mod relay;
 pub mod runtime;
