+++ crates/relayer-framework/src/full/all_for_one/mod.rs
@@ -1,4 +1,5 @@
 pub mod birelay;
+pub mod builder;
 pub mod chain;
 pub mod relay;
 pub mod runtime;
